// src/model/common/enums/ProjectStatus.ts
export enum DangerouslyProjectStatus {
    FINISHED = "016980a6-bfa2-fe8a-915b-c84551839cbf@status",
    IN_COORDINATION = "01697200-02ba-0ae8-ac44-3419b39b0914@status",
    SUSPENDED = "016976c7-c5e1-1785-9485-23b1978fb77f@status",
    IN_WORK = "01697200-0f60-3d82-bed6-968a82380a57@status",
    CLOSED = "01697200-216a-85a9-8942-b41b2e45169b@status"
}